refactor(cli): extract maintenance/setup handlers into cli_maintenance.py (ARCH-2)
Move 14 maintenance/setup handlers out of cli.py into new cli_maintenance.py
(byte-faithful bodies, no shim): _cmd_install, _install_mcp,
_get_claude_desktop_config_path, _cmd_serve, _resolve_api_token, _cmd_where,
_package_dxt, _config_encrypt_secrets, _config_set, _dashboard, _rebuild_index,
_dedup, _cleanup_gc, _status. Also moved the _ALLOWED_CONFIG_KEYS frozenset
constant (only _config_set used it).

cli.py 4272 -> 3702 LOC. With this, cli.py's only remaining top-level defs are
the orchestrator core: build_parser, _main_dispatch, main, _sync_cli_dependencies
(+ a dead _read_doi_from_frontmatter tracked separately). All domain handlers
now live in cli_* modules. cli.py is down from 8439 at ARCH-2 start (56%).

Selective re-export: cli.py re-imports 13 of 14 (all but _resolve_api_token,
which is internal — only _cmd_serve calls it). _cmd_install re-export keeps
research_hub/setup_command.py's `from research_hub.cli import _cmd_install`
working.

Forwarders (module-level + _sync_cli_dependencies): _cli_maintenance gets
get_config, ClusterRegistry, require_config, and _get_claude_desktop_config_path.
The last is an INTER-FUNCTION forwarder: _install_mcp/_cmd_install call
_get_claude_desktop_config_path() internally, and tests patch
research_hub.cli._get_claude_desktop_config_path then exercise those handlers,
so the call must observe the cli.py-level patch. Reviewer verified no recursion
(the cli.py-scoped lambda resolves to the re-exported original, never itself).

Also normalized PEP8 2-blank separators between top-level defs in the
Claude-extracted cli_paper.py + cli_maintenance.py (the extraction script
joined with 1 blank). cli_paper.py change is whitespace-only.

Extraction Claude-direct (Codex quota-blocked twice; user authorized). Tests
unchanged (pure move). Full suite 3144 passed, 0 failed.

Co-Reviewed-By: code-reviewer subagent (trigger 1: >=2 files / >=50 LOC) -> APPROVE
